Attempt to make files behave appropriately
In Firefox, right-clicking a file from the tagging page and then "Open Image in New Tab" would instead create a new tab but download the file.

This is an attempt to have the file open in the tab instead. It should still be possible to download the file with "Save Image As..."